rassus_service/requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
Django==3.0.2
django-environ==0.4.5
djangorestframework==3.11.0
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
